main: rework zoom shortcut logic
Calculate the day of year with the smallest and largest average values
instead of using hardcoded values for the minimum and maximum values
shortcuts. Calculate the padding by multiplying the supplied fraction with
the difference between the maximum and minimum value in the given zoom
interval. Account for instances where the minimum and maximum values are
the same by calculating the padding in the old manner.

Signed-off-by: Michael Yartys <michael.yartys@protonmail.com>