main: change line dash pattern and thickness
Make the dash pattern of the min/max lines denser so that rapid small
scale changes are visible. Make the lines thinner because they tended to
bleed through the edges of the lines of the yearly data. Also, change the
dashed filler line of the current year to a denser dash pattern.

Signed-off-by: Michael Yartys <michael.yartys@protonmail.com>