main: use new version of download and extract function
The download and extract function in toolkit now also takes frequency as an
argument. Change the function calls to include it. Change NH, SH, and
GLOBAL to be in lower case as the function now uses those values to
generate the URL for the data that is used to create the plots.

Signed-off-by: Michael Yartys <michael.yartys@protonmail.com>