toolkit: use requests to check whether server is up
On 2023-04-12 an issue was encountered where the thredds server hung for an
extended period of time and did not serve the data files needed for the
plotting. There is no easy way to specify how long xarray or the backend
tries to fetch the datafile before it times out. Work around this by using
the requests library to check which status code the server returns when
doing a HEAD request.

Signed-off-by: Michael Yartys <michael.yartys@protonmail.com>