monthly: add info label and remove version from title
Add a info label in the lower left corner of the plot to display the
version and source of the data used to create the plot. Also show the
latest month we have data for. In addition, remove the version number from
the title of the plot as the info label already contains the version
number.

Signed-off-by: Michael Yartys <michael.yartys@protonmail.com>