monthly: dynamically offset by month
Whenever the line that runs through all data points is visible we need to
plot the data points and trend lines with a monthly offset so that they
line up with it. Link the visibility attribute of the line to a callback
function that updates the data points and trend lines.

Signed-off-by: Michael Yartys <michael.yartys@protonmail.com>