daily: change y-range calculation
Change the y-range for the different plot shortcuts to be calculated
based on the minimum and maximum values of the data for the given
x-range of the plot shortcut selection. The start value of the y-range
takes into account the height of the text label in the lower left corner
and is offset down by this amount in order for the plot elements to not
overlap with the label. In addition, a small amount of padding is added
to the start and end of the y-range. A thick horizontal grey line is
added at y=0 in order to make this value stand out as the new y-range
calculation often results in negative y-values being included in the
range.

Signed-off-by: Michael Yartys <mail@yartys.no>