daily: use absolute min/max dates in anomaly plot
Use the dates of the absolute min/max values when plotting the yearly
min/max dots in the anomaly plot. The resulting plots have the
corresponding anomaly value for the given date as the index value in the
hovertool, and the rank is also calculated based on the absolute min/max
values.

Signed-off-by: Michael Yartys <mail@yartys.no>