Split up glyphs that go into next year
Add nan-value to the values array used to plot the forecast glyphs when it
stretches across two years. This is necessary in order to prevent the
glyph from wrapping across the screen from the end of the first year to
the start of the second. Also add fake values to other arrays.

Signed-off-by: Michael Yartys <michaely@met.no>